requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
defusedxml==0.7.1
Django==4.0.3
djangorestframework==3.13.1
drf-nested-routers==0.93.4
httpie==3.1.0
idna==3.3
multidict==6.0.2
Pygments==2.11.2
PySocks==1.7.1
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
sqlparse==0.4.2
urllib3==1.26.9
